requirements.txt
Django==5.2.4
djangorestframework==3.16.0
whitenoise==6.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
